pyproject.toml
[project]
name = "cyber-force"
version = "0.1.0"
description = "Sovereign personal agent runtime"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "pydantic>=2.11.0",
  "pyyaml>=6.0.0",
  "typer>=0.16.0",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
]

[project.scripts]
cyber-force = "cyber_force.cli:app"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

